Search_Engine_With_Langchain_Tools_And_Agents/requirements.txt
langchain
openai
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
langchain_community
faiss-cpu
streamlit
langchain-groq
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
bs4
langchain_chroma
streamlit
pymupdf
pypdf
arxiv
wikipedia
duckduckgo-search